Fix possible truncation error

+++ mistralrs-core/src/engine/mod.rs
@@ -90,7 +90,7 @@ impl Engine {
         for (logits_per_seq, seq) in zip(logits_seq, seqs.iter()) {
             let sampled = get_mut_arcmutex!(self.pipeline).sample(logits_per_seq, seq.clone());
             let next_token = handle_seq_error_stateaware!(sampled, seq);
-            let next_token_id = next_token.token as u32;
+            let next_token_id = next_token.token;
             deref_mut_refcell!(seq).add_token(next_token);
             let is_done = deref_refcell!(seq).is_done(
                 next_token_id,
@@ -105,7 +105,7 @@ impl Engine {
                 for logprob in deref_refcell!(seq).logprobs() {
                     let resp_logprob = ResponseLogprob {
                         token: handle_seq_error!(
-                            tokenizer.decode(&[logprob.token as u32], false),
+                            tokenizer.decode(&[logprob.token], false),
                             deref_refcell!(seq).responder()
                         ),
                         bytes: logprob.bytes.clone().into_bytes(),
